+++ Modules/Intent.Modules.Dapr.AspNetCore.ServiceInvocation/Templates/HttpClientConfiguration/HttpClientConfigurationTemplateRegistration.cs
@@ -27,6 +27,7 @@ namespace Intent.Modules.Dapr.AspNetCore.ServiceInvocation.Templates.HttpClientC
 
         public override string TemplateId => HttpClientConfigurationTemplate.TemplateId;
 
+        [IntentManaged(Mode.Fully)]
         public override ITemplate CreateTemplateInstance(IOutputTarget outputTarget, IList<ServiceProxyModel> model)
         {
             return new HttpClientConfigurationTemplate(outputTarget, model);