+++ Modules/Intent.Modules.Integration.HttpClients/Templates/HttpClientConfiguration/HttpClientConfigurationTemplatePartial.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using Intent.Engine;
 using Intent.Modelers.Types.ServiceProxies.Api;
+using Intent.Modules.Application.Contracts.Clients.Templates;
+using Intent.Modules.Application.Contracts.Clients.Templates.ServiceContract;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Configuration;
@@ -11,37 +13,35 @@ using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Contracts.Clients.Shared;
 using Intent.Modules.Integration.HttpClients.Shared.Templates;
+using Intent.Modules.Integration.HttpClients.Shared.Templates.HttpClientConfiguration;
+using Intent.Modules.Integration.HttpClients.Templates.HttpClient;
 using Intent.Modules.Metadata.WebApi.Models;
 using Intent.RoslynWeaver.Attributes;
-using Intent.Modules.Application.Contracts.Clients.Templates;
 using Intent.Templates;
-using Intent.Modules.Integration.HttpClients.Shared.Templates.HttpClientConfiguration;
-using Intent.Modules.Application.Contracts.Clients.Templates.ServiceContract;
-using Intent.Modules.Integration.HttpClients.Templates.HttpClient;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpTemplatePartial", Version = "1.0")]
 
 namespace Intent.Modules.Integration.HttpClients.Templates.HttpClientConfiguration
 {
-    [IntentManaged(Mode.Fully, Body = Mode.Merge)]
+    [IntentManaged(Mode.Fully, Signature = Mode.Ignore, Body = Mode.Ignore)]
     public partial class HttpClientConfigurationTemplate : HttpClientConfigurationBase
     {
         public const string TemplateId = "Intent.Integration.HttpClients.HttpClientConfiguration";
         private readonly IList<ServiceProxyModel> _typedModels;
 
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
-        public HttpClientConfigurationTemplate(IOutputTarget outputTarget, IList<ServiceProxyModel> model) 
-            : base(TemplateId, 
-                  outputTarget, 
+        public HttpClientConfigurationTemplate(IOutputTarget outputTarget, IList<ServiceProxyModel> model)
+            : base(TemplateId,
+                  outputTarget,
                   model,
                   ServiceContractTemplate.TemplateId,
                   HttpClientTemplate.TemplateId,
-                  (options, proxy, template) => 
+                  (options, proxy, template) =>
                   {
                       options.AddStatement($"http.BaseAddress = configuration.GetValue<Uri>(\"{GetConfigKey(proxy, "Uri")}\");");
                       options.AddStatement($"http.Timeout = configuration.GetValue<TimeSpan?>(\"{GetConfigKey(proxy, "Timeout")}\") ?? TimeSpan.FromSeconds(100);");
-                  }                  
+                  }
                   )
         {
             _typedModels = model;